Move google analytics import actions tests to a separate module
